+++ conda_build/domutils/conda_build_config.yaml
@@ -1,0 +1,6 @@
+python:
+    - 3.7
+    - 3.8
+    - 3.9
+    - 3.10
+    - 3.11
