+++ docs/radarTutorialIndex.rst
@@ -21,3 +21,9 @@ This tutorial demonstrates the different operations that can be performed.
 
    radarTutorial
 
+This tutorial demonstrates the different operations that can be performed.
+
+.. toctree::
+   :maxdepth: 3
+
+   radarTimeInterp
