+++ docs/test_only.py
@@ -1,7 +1,7 @@
 #comment lines to run tests on individual rst files
 
 import doctest
-doctest.testfile("legsTutorial.rst")
+#doctest.testfile("legsTutorial.rst")
 #doctest.testfile("acknowledgement.rst")
 #doctest.testfile("contribute.rst")
 #doctest.testfile("geoDoc.rst")
@@ -13,4 +13,5 @@ doctest.testfile("legsTutorial.rst")
 #doctest.testfile("radarDoc.rst")
 #doctest.testfile("radarTutorial.rst")
 #doctest.testfile("radarTutorialIndex.rst")
+doctest.testfile("radarTimeInterp.rst")
 #
