+++ domutils/radar_tools/__init__.py
@@ -9,4 +9,5 @@ from .read_fst_composite    import read_fst_composite
 from .read_stage4_composite import read_stage4_composite
 from .read_mrms             import read_mrms
 from .plot_rdpr_rdqi        import plot_rdpr_rdqi
+from .obs_process           import obs_process
 from .                      import median_filter
